rename internal types file
