+++ src/core/server/server.ts
@@ -40,7 +40,7 @@ import { mapToObject } from '../utils/';
 import { ContextService } from './context';
 import { SavedObjectsServiceSetup } from './saved_objects/saved_objects_service';
 import { RequestHandlerContext } from '.';
-import { InternalCoreSetup } from './internal_setup';
+import { InternalCoreSetup } from './internal_types';
 
 const coreId = Symbol('core');
